pyproject.toml
[tool.poetry]
name = "ssssearch"
version = "0.1.0"
description = "Static Site Semantic Search"
authors = ["Egor Machnev <egorikhelp@gmail.com>"]
readme = "README.md"
keywords = ["semantic-search", "nlp", "faiss", "llm", "search"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiofiles = "^24.1.0"
aiogram = "^3.14.0"
aiohttp = "^3.10.10"
beautifulsoup4 = "^4.12.3"
faiss-cpu = "^1.9.0"
google-generativeai = "^0.8.3"
html2text = "^2024.2.26"
nltk = "^3.9.1"
numpy = "^2.1.3"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
scikit-learn = "^1.5.2"
sentence-transformers = "^3.2.1"
tqdm = "^4.66.5"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.8"
isort = "^5.13.2"
ipython = "^8.25.0"
bandit = "^1.7.9"
pre-commit = "^3.7.1"
mypy = "^1.10.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-sugar = "^1.0.0"
pytest-cov = "^5.0.0"
pytest-async = "^0.1.1"
pytest-aio = "^1.8.1"
pytest-xdist = "^3.6.1"
polyfactory = "^2.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
exclude = [".mypy_cache", ".ruff_cache", ".venv", "venv", "build", "dist"]
line-length = 88
fix = true
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
fixable = ["ALL"]

[tool.ruff.lint.pycodestyle]
max-line-length = 120
